requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.4
coverage==7.13.4
Django==6.0.2
django-cors-headers==4.9.0
django-environ==0.13.0
django-picklefield==3.4.0
django-q2==1.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
factory_boy==3.3.3
Faker==40.5.1
gunicorn==25.1.0
idna==3.11
iniconfig==2.3.0
packaging==26.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.11.0
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.12.0
requests==2.32.5
sqlparse==0.5.5
stripe==14.4.0
typing_extensions==4.15.0
urllib3==2.6.3
whitenoise==6.12.0
